architecture-squad/requirements.txt
semantic-kernel
openai
azure-identity
python-dotenv
rich
pydantic
